Д/з разбивка на модули
