+++ js/data.js
@@ -1,0 +1,40 @@
+const SIMMILAR_OBJECTS_COUNT = 25;
+
+const NAMES = [
+  'Александр',
+  'Дмитрий',
+  'Максим',
+  'Сергей',
+  'Андрей',
+  'Алексей',
+  'Артём',
+  'Илья',
+  'Кирилл',
+  'Михаил',
+  'Никита',
+  'Матвей',
+  'Роман',
+  'Егор',
+  'Арсений',
+  'Иван',
+  'Денис',
+  'Евгений',
+  'Даниил',
+  'Тимофей',
+  'Владислав',
+  'Игорь',
+  'Владимир',
+  'Павел',
+  'Руслан',
+];
+
+const COMMENT = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export {SIMMILAR_OBJECTS_COUNT, NAMES, COMMENT};
